requirements.txt
# LAMA - Dependencies
# Install: pip install -r requirements.txt

# HTTP for price API
requests>=2.31.0

# Windows-specific (cursor tracking, window management)
# These only install on Windows
pywin32>=306; sys_platform == "win32"

# For transparent overlay
# tkinter is built-in with Python on Windows

# Dashboard (desktop app shell)
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
pywebview>=4.4.0

# Testing
pytest>=7.0
